Redesign chunk API

+++ src/amulet/level/bedrock/chunk_decode.cpp
@@ -10,8 +10,8 @@ namespace Amulet {
 
 std::unique_ptr<BedrockChunk> BedrockRawDimension::decode_chunk(
     const BedrockRawChunk& raw_chunk,
-    std::int64_t cx,
-    std::int64_t cz)
+    std::int32_t cx,
+    std::int32_t cz)
 {
     throw std::runtime_error("NotImplementedError");
 }
